Avanzar y reordenar
