+++ py_arg_reports/reporters/libro_sueldo/__init__.py
@@ -1,0 +1,1 @@
+from reporters.libro_sueldo.base import descargar_libro  # noqa
